pyproject.toml
[project]
name = "agentic-cheque-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "pillow>=10.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "pytesseract>=0.3.10",
    "streamlit>=1.40.0",
    "torch>=2.0.0",
    "transformers>=4.0.0",
    "uvicorn>=0.30.0",
]

[tool.uv]
dev-dependencies = []

[tool.setuptools.packages.find]
where = ["src"]
